Add calibration result data to window and log file
Displays calibration error for each point in Psychtoolbox window and
MATLAB command window, and logs calibration result messages to the
EyeTribe .tsv file.

When given an additional input with a conversion factor from pixels to
degrees, calibration results are in degrees of visual angle, otherwise
in pixels (or sometimes both).